+++ app/Requerimiento.php
@@ -7,6 +7,6 @@ class Requerimiento extends Model {
 	//
 	protected $primaryKey = 'idRequerimiento';
 	protected $table = 'requerimientos';
-	protected $fillable = ['idRequerimiento','idcliente','fecha_ingreso','semana','requerimiento','ingreso','idusuario','idturno'];
+	protected $fillable = ['idRequerimiento','idcliente','fecha_ingreso','semana','requerimiento','ingreso','idusuario'];
 	public $timestamps = false;
 }
