+++ composer.json
@@ -12,8 +12,7 @@
 	},
 	"require-dev": {
 		"phpunit/phpunit": "~4.0",
-		"phpspec/phpspec": "~2.1",
-		"barryvdh/laravel-dompdf": "0.6.*"
+		"phpspec/phpspec": "~2.1"
 	},
 	"autoload": {
 		"classmap": [
