+++ config/app.php
@@ -145,7 +145,6 @@ return [
 		'App\Providers\ConfigServiceProvider',
 		'App\Providers\EventServiceProvider',
 		'App\Providers\RouteServiceProvider',
-		'Barryvdh\DomPDF\ServiceProvider',
 	
 		/*
 		 *libreria de Excel
@@ -203,7 +202,6 @@ return [
 		'Form'      => 'Illuminate\Html\FormFacade',
      		'Html'      => 'Illuminate\Html\HtmlFacade',
 		'Excel'     => 'Maatwebsite\Excel\Facades\Excel',
-		'PDF' => 'Barryvdh\DomPDF\Facade',
 
 	],
