--- resources/views/PDF/invoice.blade.php
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<h3><center><strong> AUTORIZACION DE COMPRA </strong></center></h3>
-<style>
-  .clase1{
-      width: 5%;
-    }
-    
-    .clase2{
-      width: 5%;
-    }
-    
-    .textoarea1{
-    
-      width: 630px;
-    
-    }
-    
-    .textoarea2{
-    
-      width: 630px;
-    
-    }
-</style>
-</head>
-<body>
-    <br>
-
-  <table border="1">
-    
-    <tr>
-    <td class="clase1"><strong> Solicita </strong> </td>
-    <td><center> Isabel Carillo </center></td>
-        
-        <td class="clase2"><strong> Fecha </strong> </td>
-    <td><center> 07-Marzo-2017 </center></td>
-        
-    </tr>
-    
-    <tr>
-    <td class="clase1"><strong> Responsable </strong> </td>
-    <td><center> Julio Bastidas </center></td>
-        
-        <td class="clase2"><strong> Area </strong> </td>
-    <td><center> Sistemas </center></td>
-        
-    </tr>
-    
-  </table>
-    <table border="1">
-    <tr>
-      <td class="clase1"><strong> Concepto </strong> </td>
-    <td><center> Compra de Equipo de computo </center></td>
-    </tr>
-    </table>
-    
-    <br>
-    
-    <table border="1">
-    <tr>
-      <td class="textoarea1"><center><strong> Descripcion </strong>
-    </tr>
-    <tr>
-        <td class="textoarea2"><center><strong> Se compran 10 computadoras de escritorio marca Dell </strong> </center>
-    </tr>
-    </table>
-    
-      
-    <br>
-    <br>
-    
-      <center><h3>ESTIMADO DE AUTORIZACION DE COMPRA</h3></center>
-    <br>
-    <table border="1">
-    <tr>
-    
-      <td class="clase1"><strong><center> Proveedor </center></strong> </td>
-        <td class="clase1"><strong><center> Descripcion </center></strong> </td>
-        <td class="clase1"><strong><center> Precio Unitario </center></strong> </td>
-        <td class="clase1"><strong><center> Cantidad </center></strong> </td>
-        <td class="clase1"><strong><center> Total </center></strong> </td>
-        
-    </tr>
-    <tr>
-      <td><center> Office Depot </center></td>
-        <td><center> Computadoras DELL CORE 2 DUO 4GB RAM CHIP de VIdeo 512 mg con accesorios </center></td>
-        <td><center> 1000 </center></td>
-        <td><center> 10 </center></td>
-        <td><center> 10,000 </center></td>
-    </tr>
-    </table>
-
-    <br>
-    <br>
-    
-      <table border="1">
-    <tr>
-    
-      <td class="clase1"><strong><center> PREPARADO </center></strong> </td>
-        <td class="clase1"><strong><center> REVISADO </center></strong> </td>
-        <td class="clase1"><strong><center> RECIBIDO </center></strong> </td>
-        <td class="clase1"><strong><center> AUTORIZADO </center></strong> </td>
-        
-    </tr>
-    
- 
-    <tr align="center">
-      <td><center> Isabel Carillo </center></td>
-        <td><center> Julio Bastidas </center></td>
-        <td><center> Eliana </center></td>
-        <td><center> Tae Min Jung </center></td>
-    </tr>
-
-    </table>
-    
-</body>
-</html>
-
-
-
