+++ resources/views/admin/mensaje.blade.php
@@ -89,7 +89,7 @@ $( "#tipo" ).change(function() {
   	$("#usuario").prop('required', true);
   }else if ($("#tipo").val()==''){
   	$(".area").hide();
-  	$(".usuariows").hide();
+  	$(".usuarios").hide();
   }
 });
